+++ bin/deploy-gcp-function
@@ -5,6 +5,10 @@ if [ -z "$FUNCTION_NAME" ]; then
   exit
 fi
 
+if [ -z "$ENTRY_POINT" ]; then
+  ENTRY_POINT=$FUNCTION_NAME
+fi
+
 if [ -z "$PROJECT_ID" ]; then
   echo "Please provide the project ID: PROJECT_ID"
   exit
@@ -15,7 +19,7 @@ if [ -z "$GROUP_ME_BOT_ID" ]; then
   exit
 fi
 
-deploy_command="gcloud functions deploy $FUNCTION_NAME --runtime python37 --trigger-http --project $PROJECT_ID --set-env-vars GROUP_ME_BOT_ID=$GROUP_ME_BOT_ID"
+deploy_command="gcloud functions deploy $FUNCTION_NAME --entry-point $ENTRY_POINT --runtime python37 --trigger-http --project $PROJECT_ID --set-env-vars GROUP_ME_BOT_ID=$GROUP_ME_BOT_ID"
 
 echo "$deploy_command"
